requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.27.0
pytest-cov>=5.0.0
pact-python>=1.8.0 